+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Added a little more silver to surface bismuth veins, moved apatite veins closer to the surface and with deeper indicators, increased deep sapphire veins' max Y level and added a rock type, and also added a rock type to deep magnetite's veins (all overworld only) @Pyritie
 - Buffed the amount of roof tiles from clay knapping (#2178) @Pyritie
 - Added sugar cookies, toasts, jams, pickled eggs, and butter to the meal bag tag, removed banana splits (#2129) @Pyritie
+- Added assembler, smelting and heating recipes for firmalife ovens (#2220) @SpicyNoodle5
 ### Bug fixes
 - Fixed TFC bug where paths and roads can collapse on themselves when creating a staircase shape. (#2144) @Redeix
 - Fixed the HV aqueous accumulator water recipe so the amount of water can actually fit in the HV machine @Pyritie